notebooklm_auto_review/requirements.txt
playwright>=1.40.0
pyyaml>=6.0
python-docx>=1.1.0
pypdf>=4.0.0
pdfplumber>=0.10.0
loguru>=0.7.0
